pyproject.toml
[tool.poetry]
name = "livemasjid"
version = "0.1.1"
description = ""
authors = ["Yusuff Lockhat <45769466+lockhaty@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.6"
paho-mqtt = "^1.5.1"
requests = "^2.26.0"
websocket = "^0.2.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
